move Dataset class to dataset.py and change base-line model to PreAct-ResNet18
