+++ dataset.py
@@ -1,8 +1,11 @@
 import os
 import shutil
 
+import cv2
+import torch.utils.data as data
+
 def split_dataset_into_3(path_to_dataset, train_ratio, valid_ratio):
-    """
+    '''
     datasetを3つのサブセットに分割する(test,validation,train)
     
     Parameters
@@ -17,7 +20,7 @@ def split_dataset_into_3(path_to_dataset, train_ratio, valid_ratio):
     Returns
     -------
     path_to_datasetの直上にそれぞれのフォルダが作成される
-    """
+    '''
     _, sub_dirs, _ = next(iter(os.walk(path_to_dataset)))  # retrieve name of subdirectories
     sub_dir_item_cnt = [0 for i in range(len(sub_dirs))]  # list for counting items in each sub directory(class)
 
@@ -69,133 +72,47 @@ def split_dataset_into_3(path_to_dataset, train_ratio, valid_ratio):
 
     return
 
-# class ImageTransform(object):
-#     '''
-#     入力画像の前処理をするクラス
-#     train: 一定の確率で左右反転
-#            ランダムにCropしてリサイズ
-#            Tensor型に変換
-#            標準化
-# 
-#     validation、test: リサイズ
-#                       Tensor型に変換
-#                       標準化
-# 
-#     Attributes
-#     ----------
-#     data_trasnform: dict
-#         各処理をまとめておく
-#     '''
-# 
-#     def __init__(self, resize):
-#         '''
-#         Parameters
-#         ----------
-#         resize: int
-#             リサイズする画像の大きさ
-#         '''
-#         self.data_trasnform = {
-#             'train': transforms.Compose([
-#                 transforms.RandomResizedCrop(resize), 
-#                 transforms.RandomHorizontalFlip(p=0.5), # 一定の確率（ｐ）で左右反転
-#                 transforms.ToTensor(), # Tensor型に変換
-#                 transforms.Normalize([0.5, 0.5, 0.5], [0.2, 0.2, 0.2]) # 色情報の標準化
-#             ]),
-#             'validation': transforms.Compose([
-#                 transforms.Resize(256),
-#                 transforms.CenterCrop(resize),
-#                 transforms.ToTensor(),
-#                 transforms.Normalize([0.5, 0.5, 0.5], [0.2, 0.2, 0.2])
-#             ]),
-#             'test': transforms.Compose([
-#                 transforms.Resize((resize, resize)),
-#                 transforms.ToTensor(),
-#                 transforms.Normalize([0.5, 0.5, 0.5], [0.2, 0.2, 0.2])])
-#             }
-# 
-#     def __call__(self, image, phase='train'):
-#         '''
-#         Parameters
-#         ----------
-#         phase: str
-#             データセットのフェーズ
-# 
-#         Returns
-#         -------
-#         data_trasnform[phase](image): object
-#         '''
-#         return self.data_trasnform[phase](image)
-# 
-# class MyDataset(data.Dataset):
-#     '''
-#     Datasetの作成クラス
-#     PyTorchのDatasetクラスを継承させる
-# 
-#     Attrbutes
-#     ---------
-#     data_dir_path:
-#         データセットが格納されているディレクトリのパス
-#     classes: list of str
-#         ラベル名のリスト
-#     transform: object
-#         前処理クラスのインスタンス
-#     phase: str
-#         各phaseで前処理が違うため、個別に処理するフラグ
-#     data_phase: str
-#         フェーズまでが記されているパス
-#     image_paths: list of str
-#         各画像のパスのリスト
-#     '''
-# 
-#     def __init__(self, data_dir_path, classes, transform=None, phase='train'):
-#         '''
-#         Parameters
-#         ----------
-#         data_dir_path:
-#             データセットが格納されているディレクトリのパス
-#         classes: list of str
-#             ラベル名のリスト
-#         transform: object
-#             前処理クラスのインスタンス
-#         phase: str
-#             各phaseで前処理が違うため、個別に処理するフラグ
-#         '''
-#         self.data_dir_path = data_dir_path
-#         self.classes = classes
-#         self.transform = transform
-#         self.phase = phase
-# 
-#         self.data_phase = os.path.join(self.data_dir_path, self.phase)
-#         self.image_paths = [str(path) for path in Path(self.data_phase).glob("**/*.jpg")]
-# 
-#     def __len__(self):
-#         '''
-#         Returns
-#         -------
-#         画像の枚数を返す
-#         '''
-#         return len(self.image_paths)
-# 
-#     def __getitem__(self, index):
-#         '''
-#         Parameters
-#         ----------
-#         index: int
-#             各画像のインデックス
-# 
-#         Returns
-#         -------
-#         img_transformed: 
-#             前処理後の画像
-#         label: int
-#             インデックスに変換した画像ラベル
-#         '''
-#         path = self.image_paths[index]
-#         image = Image.open(path)
-# 
-#         img_transformed = self.transform(image, self.phase) # phaseに応じた前処理
-# 
-#         label = self.image_paths[index].split("/")[5] # ラベル情報を取得
-#         label = self.classes.index(label) # 取得したラベル名をインデックスに変換
-# 
-#         return img_transformed, label
+# Dataset Classの定義
+class MyDataset(data.Dataset):
+    '''
+    すでにサブセットに分割されているデータに対して使用できるDatasetクラス
+    ex: train, valid, test
+    
+    Attrbutes
+    ---------
+    file_list: list
+        画像のファイルパス
+    classes_nemes: list
+        ラベル名
+    transform: object
+        前処理
+    '''
+    def __init__(self, file_list, class_names, transform=None):
+        self.file_list = file_list
+        self.class_names = class_names
+        self.transform = transform
+
+    def __len__(self):
+        '''
+        画像の枚数を返す
+        '''
+        return len(self.file_list)
+    
+    def __getitem__(self, index):
+        '''
+        （前処理した）画像データとラベルを取得
+        '''
+        img_path = self.file_list[index]
+        image = cv2.imread(img_path)
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
+        
+        # 画像ラベルをファイル名から抜き出す
+        label = self.file_list[index].split('/')[5]
+        
+        # ラベル名を数値に変換
+        label = self.class_names.index(label)
+        
+        if self.transform is not None:
+            image = self.transform(image=image)["image"]
+        
+        return image, label
